feat:Change readme and change package name
