+++ package.json
@@ -7,7 +7,6 @@
     "dist",
     "src"
   ],
-  "types": "index.d.ts",
   "engines": {
     "node": ">=10"
   },
@@ -32,8 +31,8 @@
     "singleQuote": true,
     "trailingComma": "es5"
   },
-  "name": "tokenlist",
-  "author": "Emmanuel Antony",
+  "name": "tokendetails",
+  "author": "Omji Kushwaha",
   "module": "dist/tokenlist.esm.js",
   "size-limit": [
     {
